feat: rename vc-text-area to vc-textarea
