+++ client-app/pages/checkout/shipping.vue
@@ -2,7 +2,7 @@
   <VcSectionWidget :title="'Specify shipping details'" icon="truck"></VcSectionWidget>
 
   <VcSectionWidget :title="'Order comment'" icon="document-text">
-    <VcTextArea label="Normal input" placeholder="Please enter value" :max-length="1000" :rows="4" counter />
+    <VcTextarea label="Normal input" placeholder="Please enter value" :max-length="1000" :rows="4" counter />
   </VcSectionWidget>
 </template>
