+++ client-app/ui-kit/components/atoms/textarea/vc-textarea.vue
